PyTest DQ Framework/src/connectors/postgres/postgres_connector.py
import psycopg2
import pandas as pd

class PostgresConnectorContextManager:
    def __init__(self, db_host: str, db_name: str, db_user: str, db_password: str, db_port: int):
        self.db_host = db_host
        self.db_name = db_name
        self.db_user = db_user
        self.db_password = db_password
        self.db_port = db_port
        self.connection = None

    def __enter__(self):
        self.connection = psycopg2.connect(
            host=self.db_host,
            database=self.db_name,
            user=self.db_user,
            password=self.db_password,
            port=self.db_port
        )
        return self

    def __exit__(self, exc_type, exc_value, exc_tb):
        if self.connection:
            self.connection.close()

    def get_data_sql(self, sql):
        # exec query, result = pandas df
        """
        Execute SQL query and return result as pandas DataFrame
        """
        try:
            df = pd.read_sql(sql, self.connection)
            return df
        except Exception as e:
            raise Exception(f"Error executing SQL: {e}")


