PyTest DQ Framework/requirements.txt
# clarify version later
pytest
pandas
pyarrow
psycopg2-binary
pytest-html